Move constants to appropriate file
