+++ frontend/src/metabase/containers/DataPicker/utils.ts
@@ -1,28 +1,9 @@
-import { t } from "ttag";
-
-import { DATA_BUCKET } from "./constants";
+import {
+  DATASETS_INFO_ITEM,
+  RAW_DATA_INFO_ITEM,
+  SAVED_QUESTIONS_INFO_ITEM,
+} from "./constants";
 import { DataTypeInfoItem } from "./types";
-
-const DATASETS: DataTypeInfoItem = {
-  id: DATA_BUCKET.DATASETS,
-  icon: "model",
-  name: t`Models`,
-  description: t`The best starting place for new questions.`,
-};
-
-const RAW_DATA: DataTypeInfoItem = {
-  id: DATA_BUCKET.RAW_DATA,
-  icon: "database",
-  name: t`Raw Data`,
-  description: t`Unaltered tables in connected databases.`,
-};
-
-const SAVED_QUESTIONS: DataTypeInfoItem = {
-  id: DATA_BUCKET.SAVED_QUESTIONS,
-  name: t`Saved Questions`,
-  icon: "folder",
-  description: t`Use any question’s results to start a new question.`,
-};
 
 export function getDataTypes({
   hasModels,
@@ -36,13 +17,13 @@ export function getDataTypes({
   const dataTypes: DataTypeInfoItem[] = [];
 
   if (hasNestedQueriesEnabled && hasModels) {
-    dataTypes.push(DATASETS);
+    dataTypes.push(DATASETS_INFO_ITEM);
   }
 
-  dataTypes.push(RAW_DATA);
+  dataTypes.push(RAW_DATA_INFO_ITEM);
 
   if (hasNestedQueriesEnabled && hasSavedQuestions) {
-    dataTypes.push(SAVED_QUESTIONS);
+    dataTypes.push(SAVED_QUESTIONS_INFO_ITEM);
   }
 
   return dataTypes;
